[Microservices] Align service state with chief
